+++ StabilityMatrix.Avalonia/Views/CivArchiveBrowserPage.axaml.cs
@@ -1,0 +1,43 @@
+using System;
+using AsyncAwaitBestPractices;
+using Avalonia.Controls;
+using Avalonia.Input;
+using Injectio.Attributes;
+using StabilityMatrix.Avalonia.Controls;
+using StabilityMatrix.Avalonia.Models;
+using StabilityMatrix.Avalonia.ViewModels.CheckpointBrowser;
+
+namespace StabilityMatrix.Avalonia.Views;
+
+[RegisterSingleton<CivArchiveBrowserPage>]
+public partial class CivArchiveBrowserPage : UserControlBase
+{
+    public CivArchiveBrowserPage()
+    {
+        InitializeComponent();
+    }
+
+    private void ScrollViewer_OnScrollChanged(object? sender, ScrollChangedEventArgs e)
+    {
+        if (sender is not ScrollViewer scrollViewer)
+            return;
+
+        if (scrollViewer.Offset.Y == 0)
+            return;
+
+        var isAtEnd = Math.Abs(scrollViewer.Offset.Y - scrollViewer.ScrollBarMaximum.Y) < 1f;
+
+        if (isAtEnd && DataContext is IInfinitelyScroll scroll)
+        {
+            scroll.LoadNextPageAsync().SafeFireAndForget();
+        }
+    }
+
+    private void InputElement_OnKeyDown(object? sender, KeyEventArgs e)
+    {
+        if (e.Key == Key.Escape && DataContext is CivArchiveBrowserViewModel viewModel)
+        {
+            viewModel.ClearSearchQueryCommand.Execute(null);
+        }
+    }
+}